% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapipeline_operations.R
\name{datapipeline_validate_pipeline_definition}
\alias{datapipeline_validate_pipeline_definition}
\title{Validates the specified pipeline definition to ensure that it is well
formed and can be run without error}
\usage{
datapipeline_validate_pipeline_definition(pipelineId, pipelineObjects,
  parameterObjects, parameterValues)
}
\arguments{
\item{pipelineId}{[required] The ID of the pipeline.}

\item{pipelineObjects}{[required] The objects that define the pipeline changes to validate against the
pipeline.}

\item{parameterObjects}{The parameter objects used with the pipeline.}

\item{parameterValues}{The parameter values used with the pipeline.}
}
\description{
Validates the specified pipeline definition to ensure that it is well
formed and can be run without error.
}
\section{Request syntax}{
\preformatted{datapipeline$validate_pipeline_definition(
  pipelineId = "string",
  pipelineObjects = list(
    list(
      id = "string",
      name = "string",
      fields = list(
        list(
          key = "string",
          stringValue = "string",
          refValue = "string"
        )
      )
    )
  ),
  parameterObjects = list(
    list(
      id = "string",
      attributes = list(
        list(
          key = "string",
          stringValue = "string"
        )
      )
    )
  ),
  parameterValues = list(
    list(
      id = "string",
      stringValue = "string"
    )
  )
)
}
}

\keyword{internal}
